Rename RL event publisher module
